Added not found validation for task priority controller methods.

+++ backend-api/backend-api/Features/TaskPriority/Controller/TaskPriorityController.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using backend_api.Database.ProjectTaskRepository;
 using backend_api.Features.TaskPriority.Repository;
 using backend_api.Util;
 using Microsoft.AspNetCore.Mvc;
@@ -21,9 +22,8 @@ namespace backend_api.Features.TaskPriority.Controller
         [HttpGet("{taskId}")]
         public ActionResult<string> Get(string taskId, [FromHeader]string userId)
         {
-            // TODO
-            // Check auth/user - Unauthorised() and Forbidden() if no access to project
-            // Check task exists - NotFound()
+            // TODO: Check auth - Unauthorised()
+            // TODO: Check permissions - Forbidden()
             
             // Parse task guid and user guid
             var taskGuid = ValidationUtil.ParseGuid(taskId);
@@ -32,15 +32,23 @@ namespace backend_api.Features.TaskPriority.Controller
             {
                 return BadRequest(Errors.InvalidGuid);
             }
+            
+            // Check task exists
+            var projectTaskRepository = new ProjectTaskRepository();
+            var exists = projectTaskRepository.Exists(taskGuid);
+            if (exists.IsError)
+            {
+                return StatusCode(500, exists.ErrorException.Message);
+            }
+            if (!exists.ReturnValue)
+            {
+                return StatusCode(404, Errors.TaskNotFound);
+            }
 
+            // Read
             var result = _taskPriorityRepository.GetForTask(taskGuid);
             
-            if (result.IsError)
-            {
-                return StatusCode(500, result.ErrorException.Message);
-            }
-            
-            return StatusCode(200, result.ReturnValue);
+            return result.IsError ? StatusCode(500, result.ErrorException.Message) : StatusCode(200, result.ReturnValue);
         }
         
         
@@ -49,6 +57,9 @@ namespace backend_api.Features.TaskPriority.Controller
         [HttpPatch("{taskId}")]
         public ActionResult Patch(string taskId, [FromBody]Models.TaskPriority taskPriority, [FromHeader]string userId)
         {
+            // TODO: Check auth - Unauthorised()
+            // TODO: Check permissions - Forbidden()
+            
             // Parse task guid and user guid
             var taskGuid = ValidationUtil.ParseGuid(taskId);
             var userGuid = ValidationUtil.ParseGuid(userId);
@@ -57,24 +68,26 @@ namespace backend_api.Features.TaskPriority.Controller
                 return BadRequest(Errors.InvalidGuid);
             }
             
-            // TODO: Check valid task
+            // Check task exists
+            var projectTaskRepository = new ProjectTaskRepository();
+            var exists = projectTaskRepository.Exists(taskGuid);
+            if (exists.IsError)
+            {
+                return StatusCode(500, exists.ErrorException.Message);
+            }
+            if (!exists.ReturnValue)
+            {
+                return StatusCode(404, Errors.TaskNotFound);
+            }
             
-            if (taskPriority == null)
+            // Check priority is provided and parsed correctly
+            if (taskPriority?.getPriority() == null)
             {
                 return StatusCode(400, Errors.PriorityInvalid);
             }
 
-            var priority = taskPriority.getPriority();
-            // Check valid status
-            if (priority == null)
-            {
-                return StatusCode(400, Errors.PriorityInvalid);
-            }
-            
-            // TODO: Check User can access task
-            
-            var result = _taskPriorityRepository.Edit(taskGuid, priority);
-            
+            // Edit
+            var result = _taskPriorityRepository.Edit(taskGuid, taskPriority.getPriority());
             if (result.IsError)
             {
                 return StatusCode(500, result.ErrorException.Message);
@@ -88,10 +101,8 @@ namespace backend_api.Features.TaskPriority.Controller
         [HttpDelete("{taskId}")]
         public ActionResult Delete(string taskId, [FromHeader]string userId)
         {
-            // TODO
-            // Check auth - Unauthorised()
-            // Check permissions - Forbidden()
-            // Check task exists - NotFound()
+            // TODO: Check auth - Unauthorised()
+            // TODO: Check permissions - Forbidden()
             
             // Parse task guid and user guid
             var taskGuid = ValidationUtil.ParseGuid(taskId);
@@ -100,9 +111,21 @@ namespace backend_api.Features.TaskPriority.Controller
             {
                 return BadRequest(Errors.InvalidGuid);
             }
-
-            var result = _taskPriorityRepository.Delete(taskGuid);
             
+            // Check task exists
+            var projectTaskRepository = new ProjectTaskRepository();
+            var exists = projectTaskRepository.Exists(taskGuid);
+            if (exists.IsError)
+            {
+                return StatusCode(500, exists.ErrorException.Message);
+            }
+            if (!exists.ReturnValue)
+            {
+                return StatusCode(404, Errors.TaskNotFound);
+            }
+            
+            // Delete
+            var result = _taskPriorityRepository.Delete(taskGuid);
             if (result.IsError)
             {
                 return StatusCode(500, result.ErrorException.Message);
@@ -110,6 +133,5 @@ namespace backend_api.Features.TaskPriority.Controller
             
             return StatusCode(200);
         }
-
     }
 }
